+++ internal/servers/controller/handler.go
@@ -292,6 +292,7 @@ func wrapHandlerWithCommonFuncs(h http.Handler, c *Controller, props HandlerProp
 			// piggyback some eventing fields with the auth info proto message
 			requestInfo.EventId = info.EventId
 			requestInfo.TraceId = info.Id
+			requestInfo.ClientIp = info.ClientIp
 		} else {
 			w.WriteHeader(http.StatusInternalServerError)
 			event.WriteError(ctx, op, errors.New("unable to read event request info from context"))